move non shared code to packages/app
