+++ packages/app/package.json
@@ -67,6 +67,7 @@
     "vite-plugin-pages": "0.18.1",
     "vite-plugin-vue-layouts": "0.6.0",
     "vite-svg-loader": "3.1.2",
+    "vitest": "0.25.1",
     "vue": "3.2.31",
     "vue-i18n": "9.2.0-beta.7",
     "vue-router": "4",
