+++ packages/app/src/specs/SpecsList.vue
@@ -185,10 +185,10 @@ import { Specs_SpecsListFragment, SpecsList_GitInfoUpdatedDocument, SpecsListFra
 import { useI18n } from '@cy/i18n'
 import { buildSpecTree, fuzzySortSpecs, makeFuzzyFoundSpec, useCachedSpecs } from './spec-utils'
 import type { FuzzyFoundSpec } from './spec-utils'
-import { useCollapsibleTree } from '@packages/frontend-shared/src/composables/useCollapsibleTree'
+import { useCollapsibleTree } from './tree/useCollapsibleTree'
 import RowDirectory from './RowDirectory.vue'
 import SpecItem from './SpecItem.vue'
-import { useVirtualList } from '@packages/frontend-shared/src/composables/useVirtualList'
+import { useVirtualList } from './tree/useVirtualList'
 import NoResults from '@cy/components/NoResults.vue'
 import SpecPatternModal from '../components/SpecPatternModal.vue'
 import { useOnline, useResizeObserver } from '@vueuse/core'
