+++ packages/app/src/specs/generators/scaffold/ScaffoldGeneratorStepOne.vue
@@ -59,7 +59,7 @@ import { gql, useMutation } from '@urql/vue'
 import { ScaffoldGeneratorStepOne_ScaffoldIntegrationDocument } from '../../../generated/graphql'
 import { computed, onMounted } from 'vue'
 import { buildSpecTree } from '../../spec-utils'
-import { useCollapsibleTree } from '@packages/frontend-shared/src/composables/useCollapsibleTree'
+import { useCollapsibleTree } from '../../tree/useCollapsibleTree'
 import StandardModalFooter from '@cy/components/StandardModalFooter.vue'
 import Button from '@cy/components/Button.vue'
 import { useI18n } from '@cy/i18n'
