+++ tooling/v8-snapshot/cache/dev-linux/snapshot-meta.cache.json
@@ -48,6 +48,8 @@
     "./packages/server/lib/modes/record.js",
     "./packages/server/lib/modes/run.ts",
     "./packages/server/lib/open_project.ts",
+    "./packages/server/lib/project-base.ts",
+    "./packages/server/lib/socket-ct.ts",
     "./packages/server/node_modules/@benmalka/foxdriver/node_modules/graceful-fs/polyfills.js",
     "./packages/server/node_modules/glob/node_modules/minimatch/minimatch.js",
     "./packages/server/node_modules/graceful-fs/polyfills.js",
@@ -3527,5 +3529,5 @@
     "./tooling/v8-snapshot/cache/dev-linux/snapshot-entry.js"
   ],
   "deferredHashFile": "yarn.lock",
-  "deferredHash": "ce60d1d6cc0c77f20dcfc741105a5c53593752f814bf3cab816a3024faf68bfc"
+  "deferredHash": "6e64b278585907ecd02c6688a3964ef282453f2ca9753603ade51dc14ab6f7ee"
 }
